options-edge/backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy[asyncio]>=2.0.25
asyncpg>=0.29.0
alembic>=1.13.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
httpx>=0.26.0
anthropic>=0.40.0
langgraph>=0.2.0
langgraph-checkpoint-postgres>=2.0.0
pandas>=2.1.0
pandas-ta>=0.3.14b
numpy>=1.26.0
google-auth>=2.27.0
google-api-python-client>=2.115.0
python-dotenv>=1.0.0
